+++ android/uhabits-android/src/androidTest/java/org/isoron/uhabits/widgets/CheckmarkWidgetTest.java
@@ -48,7 +48,7 @@ public class CheckmarkWidgetTest extends BaseViewTest
     public void setUp()
     {
         super.setUp();
-        setTheme(R.style.TransparentWidgetTheme);
+        setTheme(R.style.WidgetTheme);
 
         habit = fixtures.createShortHabit();
         checkmarks = habit.getCheckmarks();