+++ android/uhabits-android/src/main/java/org/isoron/uhabits/widgets/CheckmarkWidget.kt
@@ -36,6 +36,7 @@ class CheckmarkWidget(
 
     override fun refreshData(v: View) {
         (v as CheckmarkWidgetView).apply {
+            setBackgroundAlpha(preferedBackgroundAlpha)
             setPercentage(habit.scores.todayValue.toFloat())
             setActiveColor(PaletteUtils.getColor(context, habit.color))
             setName(habit.name)