+++ android/uhabits-android/src/main/java/org/isoron/uhabits/widgets/FrequencyWidget.kt
@@ -38,6 +38,7 @@ class FrequencyWidget(
     override fun refreshData(v: View) {
         val widgetView = v as GraphWidgetView
         widgetView.setTitle(habit.name)
+        widgetView.setBackgroundAlpha(preferedBackgroundAlpha)
         (widgetView.dataView as FrequencyChart).apply {
             setColor(PaletteUtils.getColor(context, habit.color))
             setFrequency(habit.repetitions.weekdayFrequency)