+++ android/uhabits-android/src/main/java/org/isoron/uhabits/widgets/views/HabitWidgetView.java
@@ -49,6 +49,8 @@ public abstract class HabitWidgetView extends FrameLayout
 
     private StyledResources res;
 
+    private int backgroundAlpha;
+
     public HabitWidgetView(Context context)
     {
         super(context);
@@ -64,18 +66,22 @@ public abstract class HabitWidgetView extends FrameLayout
     public void setShadowAlpha(int shadowAlpha)
     {
         this.shadowAlpha = shadowAlpha;
+        rebuildBackground();
+    }
+
+    public void setBackgroundAlpha(int backgroundAlpha)
+    {
+        this.backgroundAlpha = backgroundAlpha;
+        rebuildBackground();
     }
 
     protected abstract
     @NonNull
     Integer getInnerLayoutId();
 
-    protected void rebuildBackground()
+    public void rebuildBackground()
     {
         Context context = getContext();
-
-        int backgroundAlpha =
-            (int) (255 * res.getFloat(R.attr.widgetBackgroundAlpha));
 
         int shadowRadius = (int) dpToPixels(context, 2);
         int shadowOffset = (int) dpToPixels(context, 1);